+++ src/HttpRequest.cc
@@ -193,8 +193,8 @@ std::string HttpRequest::createRequest()
   }
   if(segment_ && segment_->getLength() > 0 && 
      (request_->isPipeliningEnabled() || getStartByte() > 0)) {
-    std::string rangeHeader(fmt("bytes=%lld-",
-                                static_cast<long long int>(getStartByte())));
+    std::string rangeHeader(fmt("bytes=%" PRId64 "-",
+                                static_cast<int64_t>(getStartByte())));
     if(request_->isPipeliningEnabled()) {
       rangeHeader += util::itos(getEndByte());
     } else if(getProtocol() != Request::PROTO_FTP && endOffsetOverride_ > 0) {
