+++ src/SegmentMan.cc
@@ -120,7 +120,7 @@ SharedHandle<Segment> SegmentMan::checkoutSegment
   if(!piece) {
     return SharedHandle<Segment>();
   }
-  A2_LOG_DEBUG(fmt("Attach segment#%lu to CUID#%lld.",
+  A2_LOG_DEBUG(fmt("Attach segment#%lu to CUID#%" PRId64 ".",
                    static_cast<unsigned long>(piece->getIndex()),
                    cuid));
   piece->setUsedBySegment(true);
@@ -468,10 +468,10 @@ size_t SegmentMan::countFreePieceFrom(size_t index) const
 
 void SegmentMan::ignoreSegmentFor(const SharedHandle<FileEntry>& fileEntry)
 {
-  A2_LOG_DEBUG(fmt("ignoring segment for path=%s, offset=%lld, length=%lld",
+  A2_LOG_DEBUG(fmt("ignoring segment for path=%s, offset=%" PRId64 ", length=%" PRId64 "",
                    fileEntry->getPath().c_str(),
-                   static_cast<long long int>(fileEntry->getOffset()),
-                   static_cast<long long int>(fileEntry->getLength())));
+                   static_cast<int64_t>(fileEntry->getOffset()),
+                   static_cast<int64_t>(fileEntry->getLength())));
   ignoreBitfield_.addFilter(fileEntry->getOffset(), fileEntry->getLength());
 }
